[feat] PlayerSocketEventDelegate 에 onEventRegister 메소드 추가
