+++ PlayerSocket.podspec
@@ -8,7 +8,7 @@
 
 Pod::Spec.new do |s|
   s.name             = 'PlayerSocket'
-  s.version          = '0.1.8'
+  s.version          = '0.2.1'
   s.summary          = '플레이어 소켓입니다.'
 
 # This description is used to generate tags and improve search results.
